board model modified

+++ backend/core/models/Board.go
@@ -1,7 +1,7 @@
 package models
 
 type Board struct {
-	Id              int    `json:"id" form:"id"`
+	Id              int64  `json:"id" form:"id"`
 	Name            string `json:"boardName" form:"taskName"`
 	BackgroundColor string `json:"backgroundColor" form:"priority"`
 }
